refactor: extract registration helpers to stay under 300-line limit
Extract validateSlug, passwordStrength, PasswordStrengthBar, and
SlugStatus into registration-helpers.tsx to keep register-page.tsx
under the 300-line architecture limit (263 lines after extraction).
